+++ BreweryMaster/BreweryMaster.API/Recipe/Models/RecipeFermentingIngredientResponse.cs
@@ -8,10 +8,10 @@ namespace BreweryMaster.API.Info.Models
         public int TypeId { get; set; }
         public required string TypeName { get; set; }
         public required string Name { get; set; }
-        public float? Percentage { get; set; }
+        public decimal? Percentage { get; set; }
         public int? Extraction { get; set; }
         public int? EBC { get; set; }
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public required string Unit { get; set; }
         public string? Info { get; set; }
     }
